api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.12
langchain-openai==0.0.2
pydantic==2.6.3
langgraph==0.1.11
numpy==1.24.3
pandas==2.0.3
